mayor projecgt refactoring
