--- config/default.py
@@ -1,3 +1,0 @@
-DEBUG = True
-ORGANIZATION = "LaraTUB"
-REPO = "lara"
