--- instance/config.py.example
@@ -1,5 +1,0 @@
-GITHUB_APP_ID = 7459
-GITHUB_APP_PRIVATE_KEY = 'config/lara.private-key.pem'
-GITHUB_INSTALLATION_ID = 74063
-DIALOGFLOW_DEV_TOKEN = "Bearer <YOUR_DIALOGFLOW_DEVELOP_TOKEN>"
-DIALOGFLOW_CLIENT_TOKEN = "Bearer <YOUR_DIALOGFLOW_CLIENT_TOKEN>"
