+++ lara/__init__.py
@@ -1,15 +1,15 @@
-from flask import Flask
+from flask import Flask, g
+from lara.GithubApp import GithubApp
 
 app = Flask(__name__, instance_relative_config=True)
-
-app.config.from_object('config.default')
 app.config.from_pyfile('config.py')
 
-try:
-    app.config.from_envvar('APP_CONFIG_FILE')
-except RuntimeError:
-    pass
-
+# Connect to the Lara GithubApp
+with open(app.config['GITHUB_APP_PRIVATE_KEY'], 'r') as f:
+    private_key = f.read()
+github_app = GithubApp(app.config['GITHUB_APP_ID'], private_key)
 
 import lara.server
-import lara.webhook
+# import lara.webhook
+
+
